requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
pydantic>=2.0.0
pydantic-settings
python-dotenv==1.0.0

# AWS S3 interaction
boto3==1.34.0

# DB ORM + MySQL
sqlalchemy==2.0.40
pymysql==1.1.1

# Networking / Webhooks
requests==2.31.0

python-jose
jose  
pydantic[email]